Add notifications to Difference
On top of comments, Difference can now carry notifications. Here's how
they are different from already existing comments: comments are optional piece
of information, while notifications ought to be displayed to users, even if
there is no unified diff or details.

One canonical example is for fuzzy-matching. We always want to display
a difference, even if th only change was that the file has been renamed.
On the other hand, it's not crucial to display stderr from a command
if it didn't fail, and there as no actual difference.