Add argument completion feature to diffoscope
This should enable argument completion for diffoscope.

For argument completion `python-argcomplete` module needs to be installed.

If `python-argcomplete` module is not installed and tab is pressed then
diffoscope returns with status 1.

RangeCompleter class is used as completer for different options having
integer arguments.

Add documentation for argcomplete in README.rst.

Closes-bug: #826711